Move CLI error formatter into error reporting
