+++ packages/php-wasm/universal/src/lib/index.ts
@@ -18,8 +18,8 @@ export {
 	printDebugDetails,
 	prettyPrintFullStackTrace,
 	printResponseDebugDetails,
+	describeError,
 } from './error-reporting';
-export { describeError } from './error-formatting';
 export { UnhandledRejectionsTarget } from './wasm-error-reporting';
 export { HttpCookieStore } from './http-cookie-store';
 export type { IteratePhpFilesOptions as IterateFilesOptions } from './iterate-files';
