+++ packages/playground/cli/tests/run-cli.spec.ts
@@ -8,7 +8,6 @@ import {
 	resolveWorkerCount,
 } from '../src/run-cli';
 import type { RunCLIArgs, RunCLIServer } from '../src/run-cli';
-import { describeError } from '@php-wasm/universal';
 import type { MockInstance } from 'vitest';
 import { vi } from 'vitest';
 import { mkdtemp, writeFile } from 'node:fs/promises';
@@ -1198,54 +1197,6 @@ describe('start command', () => {
 			rmSync(tmpDir, { recursive: true, force: true });
 		}
 	}, 180000);
-});
-
-describe('describeError', () => {
-	test('falls back to Error cause when message is empty', () => {
-		const error = new Error('', {
-			cause: new Error(
-				'Error when executing the blueprint step #1: PHP.run() failed with exit code 255.'
-			),
-		});
-
-		const description = describeError(error);
-		expect(description).toContain(
-			'Error when executing the blueprint step #1'
-		);
-		expect(description.startsWith('Error —')).toBe(false);
-	});
-
-	test('terminates circular cause chains', () => {
-		const error = new Error('');
-		error.cause = error;
-
-		expect(describeError(error)).toContain('[Circular error cause]');
-	});
-
-	test('preserves empty-message formatting through nested causes', () => {
-		const error = new Error('', {
-			cause: new Error('', {
-				cause: new Error('Inner failure'),
-			}),
-		});
-
-		const description = describeError(error);
-		expect(description).toContain('Inner failure');
-		expect(description).not.toContain('Error — caused by:');
-	});
-
-	test('describes local fields before cause', () => {
-		const description = describeError({
-			name: 'ErrnoError',
-			errno: 20,
-			code: 'ENOTDIR',
-			cause: new Error('Inner failure'),
-		});
-
-		expect(description).toBe(
-			'ErrnoError — errno: 20 — code: ENOTDIR — caused by: Inner failure'
-		);
-	});
 });
 
 describe('php command', () => {
